started modifying lattice sequence
