+++ python_prototype/qmcpy/discrete_distribution/__init__.py
@@ -1,2 +1,3 @@
 # API
-from .distributions import *
+from .iid_generators import IIDStdUniform, IIDStdGaussian
+from .lds_generators import Lattice, Sobol
