+++ python_prototype/qmcpy/discrete_distribution/mps_refactor/__init__.py
@@ -1,0 +1,2 @@
+from .digital_sequence import DigitalSeq
+from .lattice_sequence import LatticeSeq
